Add "required items" to immersive questing
The primary purpose of this change is to add the required items for a
quest to the quest dialog display at the quest completion step. Because
this involved a rework of the immersive questing state machine, a few
other fixes were also implemented.

Progress quest dialog is now available, closing quest windows is
more consistent, auto-quests should now be ignored properly, there
should be no (or at least fewer) instances of stale/incorrect quest data
not being reset between dialogs, fewer clicks are required when the
final quest line is displayed, and quest text parsing is improved.