Enhance animation and auto-quest behaviors
Revamped the the PlayerModel frames to use more dynamic animations and
to increase the model size a bit. Also added compensating code for when
the auto-camera doesn't quite get it right.

Some other fixes include better behavior with automatic/pop-up quests, a
placeholder model instead of "talking to yourself", and some fixes to
reward and required items displays.