Change micro buttons to use OnClick events
Updated the micro buttons to use OnClick handlers instead of OnMouseDown
events. There are some odd edge cases where the event gets stuck and the
frame toggling blows up the stack. Since the default frames only use
OnMouseDown in order to set a "clicked" visible state, which we don't
have, this seems like the simplest way to avoid that.