First pass at fixing some objective tracker issues
This is an initial pass at fixing a variety of problems with the
objective tracker. This commit should fix the performance problems
encountered when using the quest tracker and the compass especially. The
compass behavior has been altered slightly to work only with "local"
quests (as defined by the active map, or current map by default). Corpse
tracking has also been fixed.