Touch all the things!
This is first pass of a big refactor for 8x line. This pass has two
goals:

1 - update everything to use prettification/format as defined by VS code
and the vscode-lua extension

2 - pull nearly all script tags out of XML and inline to the Lua, as has
been done previously on a case-by-case basis already

This is in prep for a remodularization pass and loading change, since
the current loading methodology does not work well for 8x with the new
UI. I figured this was a good time to touch everything and break history
given those upcoming goals and 8x testing requirements in general.