Multiple fixes to the character window manager
The system for showing/hiding tabs in the character windows has been
updated to fix a number of issues. The keybindings no longer override
existing bindings, updated bindings are noticed immediately instead of
only on load, and the secure execution path in general has been cleaned
up quite a bit.

Also updated the changelog with these and earlier fixes.