Update PvP indicator behaviors
The PvP indicators are now less visible in the steady state. After
initial flag they are more visible for a few seconds. On mouse over for
the health globe now highlights the PvP indicator and provides a tooltip
with your PvP status (why flagged and how much time remaining if
applicable).