Move currency details to character tab (fix #43)
Sounds innocuous but this change ended up being fairly large as I ripped
the character menu system out of all the paperdoll stuff too to make it
reusable, which ended up being a pretty significant refactor.

The currency tab has a spot for raid lockout info that I hope to be able
to add. That's iffy for 8x though I might have to hide that out and just
leave the first currency iteration as a really wide window.

Oh and clicking the currency button on the bag in combat does not work
currently (silently does nothing); I need to look at making the entire
bag frame a secure handler though to ultimately address that (which I
intend to do to address other issues anyway).