Spellbook supports applying and removing glyphs
This update gives glyph support to the spellbook. This assumes glyphs
will only ever apply to spelltab "2", which I think is true. The visuals
are just the defaults so they could probably use some sprucing up.

Also included in this change is the ability to open directly to
individual spellbook tabs via the windowpanelopen attribute, and more
advanced toggle vs open-only logic.