Allow minimap to anchor at top
This change adds #60, a setting to anchor the minimap at the top or
bottom of the screen. In the future I'd like to allow full custom
minimap placement with corner anchors as "set" options.

This also updates the way the MinimapCluster is setup, which fixes
issues with things that anchor to that (like the default tracker,
vehicle seat indicator, etc.) and should fix some issues there whether
or not the custom quest tracker is in use. Fixes #55.