Avoid doing some things when not "in world"
This is an attempt to address the recent bug where zoning/loading into
the world sometimes results in a "hanging" loading screen (no progress
bar) followed by a completely broken/not shown UI. We set a flag on
leaving/entering world and update several event handlers to check for
that flag before trying to do stuff. These are based on events that
appear to fire between the leaving and entering world events (using
/etrace).

It's possible this issue occurs even earlier in the pipleine, since
LOADING_SCREEN_ENABLED fires first, then several other events, then
PLAYER_LEAVING_WORLD. One of those events might also be causing issues
that would not be addressed by this fix attempt. If that is the case the
only one we do stuff with that I've seen in an etrace is
MIRROR_TIMER_STOP, but that handler looks ultra simple and innocuous.