Use default microbuttons to avoid secure taint
Our custom micro buttons have always been a bit glitchy, but prior to
8.2 it wasn't anything severe enough to worry about. With the change in
8.2 to make ShowUIPanel and HideUIPanel protected calls, suddenly most
of our custom micro buttons simply don't work in combat. And out of
combat they have the potential to taint large portions of frame position
stuff.

Since it is basically impossible to reproduce the default micro button
functionality in the restricted environment, we instead *keep* the
default micro buttons created by the BlizzUI and just reskin them (along
with a bunch of other stuff) to get them to look, feel, and behave the
way we want them to.