Part 1 of the aura overhaul
This is the first part of the aura overhaul. The only auras directly
affected by this work are the ones in the player's personal aurabar.
Significant work still needs to be done to optimize auras for other
unitframes.

For various reasons this ends up touching a lot of things with frame
management, scaling, moving, etc. This change also introduces the basis
for a new layout manager which will handle all of the dynamic
positioning of the mainbar elements (eventually, not done yet). The
upside of that work is removal of a bunch more global frame names.

At minimum certain strings need to be localized and changelog updated
before this could release "as-is", and ideally more work should happen
before a release at all. But I'm committing this stuff since the player
auras work and support the new legacy and secure types, and could use
testing.