Update Prot Warrior dodge to Charge
In 9.0.1 they removed "Intercept" from prot and gave a nerfed version to all warriors called "Intervene" that only has 1 charge. This changes it so prot uses charge as the other specs would.

Alternatively, you could change the spell ID to 3411 (either for all specs or just prot) and it'll use the "new" Intervene instead.